manifest: move Parameter into a file of its own
